+++ cmd/frontend/internal/app/pkg/updatecheck/client.go
@@ -100,11 +100,7 @@ func getAndMarshalSiteActivityJSON(ctx context.Context, criticalOnly bool) (json
 	if err != nil {
 		return nil, err
 	}
-	contents, err := json.Marshal(siteActivity)
-	if err != nil {
-		return nil, err
-	}
-	return json.RawMessage(contents), nil
+	return json.Marshal(siteActivity)
 }
 
 func hasSearchOccurred(ctx context.Context) (bool, error) {
@@ -150,11 +146,7 @@ func getAndMarshalCampaignsUsageJSON(ctx context.Context) (json.RawMessage, erro
 	if err != nil {
 		return nil, err
 	}
-	contents, err := json.Marshal(campaignsUsage)
-	if err != nil {
-		return nil, err
-	}
-	return json.RawMessage(contents), nil
+	return json.Marshal(campaignsUsage)
 }
 
 func getAndMarshalCodeIntelUsageJSON(ctx context.Context) (json.RawMessage, error) {
@@ -171,11 +163,7 @@ func getAndMarshalCodeIntelUsageJSON(ctx context.Context) (json.RawMessage, erro
 	if err != nil {
 		return nil, err
 	}
-	contents, err := json.Marshal(codeIntelUsage)
-	if err != nil {
-		return nil, err
-	}
-	return json.RawMessage(contents), nil
+	return json.Marshal(codeIntelUsage)
 }
 
 func getAndMarshalSearchUsageJSON(ctx context.Context) (json.RawMessage, error) {
@@ -191,11 +179,7 @@ func getAndMarshalSearchUsageJSON(ctx context.Context) (json.RawMessage, error) 
 	if err != nil {
 		return nil, err
 	}
-	contents, err := json.Marshal(searchUsage)
-	if err != nil {
-		return nil, err
-	}
-	return json.RawMessage(contents), nil
+	return json.Marshal(searchUsage)
 }
 
 func updateURL(ctx context.Context) string {