+++ dev/licenses-npm.sh
@@ -9,8 +9,4 @@ FAIL_ON='UNKNOWN;GPL-1.0-only;GPL-1.0-or-later;GPL-2.0-only;GPL-2.0-or-later;GPL
 {
   # Webapp, native integrations and browser extension
   ./node_modules/.bin/license-checker --production --csv --failOn "$FAIL_ON"
-  # LSIF
-  pushd cmd/precise-code-intel >/dev/null
-  ../../node_modules/.bin/license-checker --production --csv --failOn "$FAIL_ON" | tail -n +2
-  popd >/dev/null
 } | uniq >ThirdPartyLicensesNpm.csv