+++ web/src/search/input/toggles/Toggles.tsx
@@ -3,7 +3,7 @@ import * as H from 'history'
 import RegexIcon from 'mdi-react/RegexIcon'
 import classNames from 'classnames'
 import FormatLetterCaseIcon from 'mdi-react/FormatLetterCaseIcon'
-import { PatternTypeProps, CaseSensitivityProps, InteractiveSearchProps } from '../..'
+import { PatternTypeProps, CaseSensitivityProps, InteractiveSearchProps, CopyQueryButtonProps } from '../..'
 import { SettingsCascadeProps } from '../../../../../shared/src/settings/settings'
 import { SearchPatternType } from '../../../../../shared/src/graphql/schema'
 import { isEmpty } from 'lodash'
@@ -11,12 +11,15 @@ import { submitSearch } from '../../helpers'
 import { QueryInputToggle } from './QueryInputToggle'
 import { isErrorLike } from '../../../../../shared/src/util/errors'
 import CodeBracketsIcon from 'mdi-react/CodeBracketsIcon'
+import { generateFiltersQuery } from '../../../../../shared/src/util/url'
+import { CopyQueryButton } from './CopyQueryButton'
 import { VersionContextProps } from '../../../../../shared/src/search/util'
 
 export interface TogglesProps
     extends PatternTypeProps,
         CaseSensitivityProps,
         SettingsCascadeProps,
+        CopyQueryButtonProps,
         Partial<Pick<InteractiveSearchProps, 'filtersInQuery'>>,
         VersionContextProps {
     navbarSearchQuery: string
@@ -92,8 +95,23 @@ export const Toggles: React.FunctionComponent<TogglesProps> = (props: TogglesPro
         submitOnToggle({ newPatternType })
     }
 
+    const fullQuery = [
+        props.navbarSearchQuery,
+        props.filtersInQuery && generateFiltersQuery(props.filtersInQuery),
+        `patternType:${props.patternType}`,
+        props.caseSensitive ? 'case:yes' : '',
+    ]
+        .filter(queryPart => !!queryPart)
+        .join(' ')
+
     return (
         <div className={classNames('toggle-container', props.className)}>
+            {props.copyQueryButton && (
+                <CopyQueryButton
+                    fullQuery={fullQuery}
+                    className="toggle-container__toggle toggle-container__copy-query-button"
+                />
+            )}
             <QueryInputToggle
                 {...props}
                 title="Case sensitivity"