+++ web/src/views/ViewContent.tsx
@@ -5,7 +5,7 @@ import { MarkupKind } from '@sourcegraph/extension-api-classes'
 import H from 'history'
 import { QueryInputInViewContent } from './QueryInputInViewContent'
 import { View, MarkupContent } from 'sourcegraph'
-import { CaseSensitivityProps, PatternTypeProps } from '../search'
+import { CaseSensitivityProps, PatternTypeProps, CopyQueryButtonProps } from '../search'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { hasProperty } from '../../../shared/src/util/types'
 import { isObject } from 'lodash'
@@ -19,6 +19,7 @@ export interface ViewContentProps
     extends SettingsCascadeProps,
         PatternTypeProps,
         CaseSensitivityProps,
+        CopyQueryButtonProps,
         VersionContextProps {
     viewContent: View['content']
     location: H.Location